Moving some widgets to a new path and updating imports
