+++ lib/modules/home/home_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:wallet/modules/home/home_controller.dart';
-import 'package:wallet/modules/shared/widgets/account.dart';
+import 'package:wallet/modules/shared/widgets/fragments/account.dart';
 import 'package:wallet/modules/shared/widgets/header.dart';
 
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
