+++ lib/modules/shared/widgets/footer.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:wallet/core/constants/theme/app_theme.dart';
-import 'package:wallet/modules/shared/widgets/button.dart';
+import 'package:wallet/modules/shared/widgets/fragments/button.dart';
 
 // ignore: must_be_immutable
 class CFooter extends StatelessWidget {
