+++ lib/modules/shared/widgets/menu.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:wallet/core/constants/app_images.dart';
 import 'package:wallet/core/constants/theme/app_theme.dart';
-import 'package:wallet/modules/shared/widgets/menu_option.dart';
+import 'package:wallet/modules/shared/widgets/fragments/menu_option.dart';
 
 class Menu extends Drawer {
   final GlobalKey<ScaffoldState> _scaffoldKey = GlobalKey();
